Add input key to result object
When running PDFBarcodeJS on multiple files concurrently and a callback occurs, there is no convenient way to identify which input source the result object applies to.  To solve this, a new key (input) is added to the result object based on the the input value passed to PDFBarcodeJS's decode function when called.